:fire: remove: 불필요 코드 삭제
