+++ srcs/Client/srcs/Thread.cpp
@@ -1,4 +1,5 @@
 #include "Thread.hpp"
+#include <sstream>
 
 std::thread *Thread::thread_arr[MAX_THREAD] = {nullptr};
 
@@ -9,12 +10,16 @@ void Thread::cleanThread()
     {
         if (thread_arr[i] == nullptr)
             continue;
-        std::cout << "thread_arr[" << i << "]->join()..." << std::endl;
+        std::stringstream ss;
+        ss << "thread_arr[" << i << "]->join()...";
+        std::cout << ss.str() << std::endl;
         // thread_arr[i]->join();
         thread_arr[i]->detach();
         delete thread_arr[i];
         thread_arr[i] = nullptr;
-        std::cout << "thread_arr[" << i << "]->join()... done" << std::endl;
+        ss.clear();
+        ss << "thread_arr[" << i << "]->join()... done";
+        std::cout << ss.str() << std::endl;
     }
     std::cout << "cleanThread over..." << std::endl;
 }
