+++ srcs/Client/srcs/rendering.cpp
@@ -1,23 +1,64 @@
 #include <iostream>
-#include <string>
-#include <thread>
+
 #ifdef __APPLE__
 #include "../../../mac_opengl/include/GLFW/glfw3.h"
 #else
 #include <GLFW/glfw3.h>
 #endif
+
+#include "Client.hpp"
 #include "GameFrame.hpp"
 #include "Cache.hpp"
-
-#define WIDTH 1500
-#define HEIGHT 750
-
-using namespace std;
+#include "Util.hpp"
 
 void singlePlay();
 void multiPlay();
 
-void key_callback(GLFWwindow *window, int key, int scancode, int action, int mods);
+void key_callback(GLFWwindow *window, int key, int scancode, int action, int mods)
+{
+    Client *client = Client::getInstance();
+    (void)scancode;
+    (void)mods;
+    if (action == GLFW_PRESS)
+    {
+        switch (key)
+        {
+        case GLFW_KEY_W:
+        case GLFW_KEY_UP:
+            std::cout << "player = UP" << std::endl;
+            client->sendPaddleDirection(UP);
+            break;
+        case GLFW_KEY_S:
+        case GLFW_KEY_DOWN:
+            std::cout << "player = DOWN" << std::endl;
+            client->sendPaddleDirection(DOWN);
+            break;
+        case GLFW_KEY_R:
+            // resetKey();
+            break;
+        case GLFW_KEY_ESCAPE:
+            std::cout << "ESC key pressed" << std::endl;
+            glfwSetWindowShouldClose(window, GLFW_TRUE);
+            Cache::atom_stop = true;
+            // cleanMemory();
+            // exit(EXIT_SUCCESS);
+            break;
+        default:
+            std::cout << "Other key pressed" << std::endl;
+            break;
+        }
+    }
+    else if (action == GLFW_RELEASE)
+    {
+        switch (key)
+        {
+        case GLFW_KEY_W:
+            break;
+        case GLFW_KEY_S:
+            break;
+        }
+    }
+}
 
 void error_callback(int error, const char *description)
 {
@@ -27,7 +68,7 @@ void error_callback(int error, const char *description)
 void framebuffer_size_callback(GLFWwindow *window, int width, int height)
 {
     (void)window;
-    cout << "framebuffer_size_callback" << endl;
+    std::cout << "framebuffer_size_callback" << std::endl;
     // Viewport를 새 윈도우 크기에 맞게 조정합니다.
     glViewport(0, 0, width, height);
 
@@ -63,13 +104,13 @@ static void mouse_button_callback(GLFWwindow *window, int button, int action, in
         switch (button)
         {
         case GLFW_MOUSE_BUTTON_LEFT:
-            cout << "Left button pressed" << endl;
+            std::cout << "Left button pressed" << std::endl;
             break;
         case GLFW_MOUSE_BUTTON_RIGHT:
-            cout << "Right button pressed" << endl;
+            std::cout << "Right button pressed" << std::endl;
             break;
         default:
-            cout << "Other button pressed" << endl;
+            std::cout << "Other button pressed" << std::endl;
             break;
         }
         break;
@@ -77,13 +118,13 @@ static void mouse_button_callback(GLFWwindow *window, int button, int action, in
         switch (button)
         {
         case GLFW_MOUSE_BUTTON_LEFT:
-            cout << "Left button released" << endl;
+            std::cout << "Left button released" << std::endl;
             break;
         case GLFW_MOUSE_BUTTON_RIGHT:
-            cout << "Right button released" << endl;
+            std::cout << "Right button released" << std::endl;
             break;
         default:
-            cout << "Other button released" << endl;
+            std::cout << "Other button released" << std::endl;
             break;
         }
         break;
@@ -96,7 +137,7 @@ static void mouse_button_callback(GLFWwindow *window, int button, int action, in
 static void cursor_position_callback(GLFWwindow *window, double xpos, double ypos)
 {
     (void)window;
-    cout << "Cursor Position at (" << xpos << " : " << ypos << ")" << endl;
+    std::cout << "Cursor Position at (" << xpos << " : " << ypos << ")" << std::endl;
 }
 
 void rendering()
@@ -105,15 +146,13 @@ void rendering()
     glfwSetErrorCallback(error_callback);
 
     if (!glfwInit())
-    {
-        exit(EXIT_FAILURE);
-    }
+        error_msg(FATAL_ERR);
 
-    GLFWwindow *window = glfwCreateWindow(WIDTH, HEIGHT, "Hello World", NULL, NULL);
+    GLFWwindow *window = glfwCreateWindow(WIDTH, HEIGHT, "pong", NULL, NULL);
     if (!window)
     {
         glfwTerminate();
-        exit(EXIT_FAILURE);
+        error_msg(FATAL_ERR);
     }
 
     glfwMakeContextCurrent(window);
